refactor `RewardGrid` into folders
