+++ packages/ui-core/src/base/RewardGrid/Label/index.module.scss
@@ -1,0 +1,6 @@
+// Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+.label {
+  color: var(--text-color-secondary);
+}
