+++ packages/ui-core/src/base/RewardGrid/Root/index.module.scss
@@ -1,0 +1,7 @@
+// Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+.root {
+  margin-top: 0.5rem;
+  padding: 0 0.5rem;
+}
