+++ packages/ui-core/src/base/RewardGrid/Row/index.module.scss
@@ -1,0 +1,14 @@
+// Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+.body {
+  border-bottom: 1px solid var(--border-primary-color);
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  padding: 1.05rem 0;
+
+  &:last-child {
+    border-bottom: none;
+    padding-bottom: 0.5rem;
+  }
+}
