+++ packages/ui-core/src/base/RewardGrid/common.module.scss
@@ -1,0 +1,6 @@
+// Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+.row {
+  padding: 0.6rem 0;
+}
